requirements.txt
python-dotenv==1.0.0
requests==2.31.0
websockets==12.0
py-clob-client==0.18.0
eth-account==0.10.0
web3==6.15.1
aiohttp==3.9.3
asyncio-throttle==1.0.2
